# server/swagger.yaml
...
openapi: 3.0.0
info:
  title: Man
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /face/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - Face
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - Face
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Face"
    post:
      summary: CRUD ACTION update
      tags:
        - Face
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Face"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Face"
    
  #-----  User  -----
...
  /ears/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - ears
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - ears
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ears"
    post:
      summary: CRUD ACTION update
      tags:
        - ears
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ears"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ears"
  /ears/findByears/{key}:
    get:
      summary: CRUD ACTION findByears
      tags:
        - ears
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ears"
    
  #-----  eyes  -----
...
  /eyes/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - eyes
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - eyes
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/eyes"
    post:
      summary: CRUD ACTION update
      tags:
        - eyes
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/eyes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/eyes"
  /eyes/findByeyes/{key}:
    get:
      summary: CRUD ACTION findByeyes
      tags:
        - eyes
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/eyes"
    
  #-----  man  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Face:
      properties:
        _id:
          type: string
        ears:
          type: number
        eyesnum:
          type: number
        mouth:
          type: boolean
        nose:
          type: boolean
      required:
        -  _id
...
    ears:
      properties:
        _id:
          type: string
        desc:
          type: string
        type:
          type: string
      required:
        -  _id
    eyes:
      properties:
        _id:
          type: string
        color:
          type: string
        type:
          type: object
      required:
        -  _id
...